feat: pipeline hardening + fp16 default + multithread + 4-way harness
Pipeline (src/)
- chunked inference for inputs > 512 tokens (sliding window + dedupe);
  fixes silent PII loss past chunk boundary
- tunable transition biases (TransitionBiases: background, enterSpan,
  continueSpan) added to constrained Viterbi — upstream-recommended
  precision/recall lever
- forward-backward marginals replace mean-of-softmax for span scoring;
  posterior-calibrated confidences
- per-category thresholds (categoryThresholds map) on top of global
- recognizer-vs-recognizer dedupe via shared dedupeSpans logic
- ORT session multithread support (intraOpNumThreads, interOpNumThreads)

Defaults
- DEFAULT_VARIANT for ModelManager and CPU backend: int4f16/fp32 → fp16
  (F1-equivalent to fp32, ~17% faster than int8 on Apple M-series)
- eval default variant: int8 → fp16
- thread sweep on M5 Pro: ORT auto already optimal (31.7 ms/sample);
  knob exposed for callers wanting to cap CPU

CLI
- nullpii serve --enter-bias / --background-bias / --continue-bias
- nullpii serve --threshold N
- nullpii serve --threads N

Eval framework (packages/eval/)
- en-templates.txt: +22 realistic Claude Code prompts (debug sessions,
  code review, RFC, customer support, CI/incident, secret leaks). Now
  62 bundled-en samples.
- long-prompts-en-baseline.jsonl: 62 dev prompts ~3.1k chars each, PII
  positioned past chunk boundary. Proves chunking value (F1: 0.0 →
  0.571 on this dataset, Δ=+0.571).
- run_sweep.py + eval_best_config.py: bias/threshold sweep harness
- eval_chunking_proof.py: chunking on/off comparison
- eval_backend_check.py: cpu/mps × variant matrix
- eval_4way.py: nullpii vs OpenAI bare HF pipeline vs Presidio vs spaCy
- openai_pipeline_predictor with auto MPS device + batch_size
- nullpii readiness probe drains stderr (CoreML logs no longer break
  startup)

Headline numbers
- bundled-en: nullpii 0.751 vs OpenAI bare HF 0.514 (Δ=+0.236)
- 9-dataset avg Δ(nullpii − OpenAI bare HF): +0.229 — confirms our
  Viterbi/posterior decoder adds material value over upstream pipeline
- bias sweep best on bundled-en: continueBias=0.5 → 0.762 (+0.011);
  generalization across locales: avg Δ ≈ 0 (per-locale optima differ)

Docs (READ ME, docs/index.md, docs/guide/{backends,eval-results,
getting-started,how-it-works,vs-presidio}.md)
- variant tables refreshed: fp16 highlighted as default
- M5 Pro 48GB explicit in eval-results methodology
- MPS coverage caveat (~24/365 ops CoreML-eligible) documented
- vs-presidio feature matrix: OpenAI bare HF column added

+++ packages/eval/src/nullpii_eval/run_accuracy.py
@@ -25,7 +25,7 @@ from .metrics import evaluate, macro_f1
 )
 @click.option("--max-samples", type=int, default=200)
 @click.option("--backend", default="cpu")
-@click.option("--variant", default="int8")
+@click.option("--variant", default="fp16")
 def main(out: str, dataset: str, max_samples: int, backend: str, variant: str) -> None:
     predictor = nullpii_predictor(backend=backend, variant=variant)
     if dataset == "bundled":
